Replace the uaago with go-cfclient
We'd like to support a wider range of CF configuration. The uaago
library has been somewhat limited in a way of providing the
authentication as it has been limiting the `grant_type` to
`client_credential`.

We have found the need to implement the authentication based on the
username and password instead, to provide our tenants with the ability
to manually deploy the nozzle for themselves.

We no longer require the doppler url as we obtain the API endpoint
from the new client library.