Update dependencies
Some of the libraries were very old, particulary the `noaa` library. The
implementation of the library didn't work for us, as it was focused
around a single consumer, which did export metrics from all applications
at first but then it has overwritten an internal variable each time,
making it exporting the last application found instead. This isn't the
ideal case and upgrading the library may result in positive behaviour of
the metric exporter.