Remove unsupported metric processors
The very basic aim of this application, is to obtain the cpu, memory and
disk usage of each other application that the provided user has an
access to. We don't need these extra metrics at the current time,
allowing us to delete it from the repo for the time being.