Refactor the application
The NOAA library has been previously updated. This means the
graphite-nozzle application may no longer work as expected. We're
working our way around in order to achieve the basic use case.

We're implementing the metricTemplate system, which will allow our
tenants to customise their metric naming to fit their needs.

We're now streaming a struct as oppose to an event, that consists more
information which then can be used to our advantage.

The firehose usage and the prefixName have been deprecated and removed
from this application. Sorry.

We're adjusting the tests, to cover more as well as fixing some of the
broken usage due to the refactoring process.