Write up the tests to support refreshToken actions
We'd like to make sure, our loop is working correctly and when the
refresh token actually expires, it attempts to log in once again.

We're covering the grant_type checkup as well as the mocked outcome from
the endpoints.

In the test we must respond with a JSON response that includes the
expire timeout in seconds as `expires` or `expires_in`. But oauth.Token
does not have such json enconding.

The oauth code would parse the token object as described here[1].
The json must follow the format defined by this struct [2].
As the stuct is in an internal module, we must define it again.

[1] https://github.com/golang/oauth2/blob/master/internal/token.go#L208-L220
[2] https://github.com/golang/oauth2/blob/master/internal/token.go#L56-L65