Improve token refreshment
We have implemented a new struct in order to share some values between
our functions as oppose to use a lot of dependencies. This has been
mainly reasoned, to support a `.RefreshTokenFrom()` functionality
provided by the `go-cfclient` library.

We have made use of the mentioned functionality to improve the way we're
handling the token refreshment for each consumer.

The tests have been adjusted, to reflect the newly implemented structure
of the code.