Only GUIDs are needed, so request less app info
The existing .ListApps() query[1] has `inline-relations-depth` set to
`2`, which causes it to recurse twice into each listed relation and
fetch more information to be included in the response.

This query is very expensive (especially on large orgs) and due to the
runtime it seems to be especially prone to hitting race conditions with
apps being added and removed while the query is running, which causes
API errors to be generated by the cloud controller.

We only actually need the GUIDs for the returned applications, so by
switching to .ListAppsByQuery() and not setting `inline-relations-depth`
the querytime is reduced to around ~2s, down from ~10-20s. Hopefully
this should also reduce the load on the cloud controller.

[1] https://github.com/cloudfoundry-community/go-cfclient/blob/master/apps.go#L217