Properly retrieve space and org names of the app
Retrieving them through SpaceData property resulted in empty strings
which resulted in breaking our dashboards as they were based in a
per-space filtering.

Also, it attempted to extract metrics from all kinds of events,
resulting in a large number of 0 valued metrics.

In tests we use real GUIDs for space and org so that they are not
confused with the application GUIDs and the API endpoints look like real
requests.

The `app` variable in testSpaceResource and testOrgResource wasn't being
used so I'm passing in the new GUID strings instead.