Fix yaml example in README
If you don't quote the METRIC_TEMPLATE value, then the cf cli yaml
parser gets confused by the curly brackets and tries to parse them as
a yaml object.  This results in the error message "did not find
expected key" which is not super helpful :(

This was fixed by quoting the string.  I'm updating the README to make
this clearer for future users.