Keep App data up to date each UPDATE_FREQUENCY
Previously if an app was updated after a metrics stream was started, any
updates to the app would not be reflected in the output metrics.
For example; a `cf rename` call would update an application name, but
the output metrics from this connector would still be labeled with the
old app name unless it was restarted or the app was stopped for enough
time for the stream to get stopped and reconected.

This change refactors the stream handling routines to allow recieving
udpated instances of the app data during each UPDATE_FREQUENCY period.
This will keep not just the name, but any other application, space or
org data update to date.

This change also surfaced some race conditions (mostly in the tests)
that needed to be addressed.