Refactor to a new events package
The pre-existing structure was odd. For instance, `metricProcessor`
did not fulfil the `Processor` interface and was in the `main`
package. `metrics.Stream` was a value put into a channel rather than
resembling a stream of data.

This (large) commit does the minimal work necessary to refactor out a
new Events package. The result is somewhat more natural, but there may
be some worthwhile remaining improvements to tests.

Storing the template in `ContainerMetricProcessor` is not necessarily
the best choice, and was chosen merely for convenience in an already
major change.