Fix for Travis not running the tests
A Travis build [1] passed despite containing failing tests, because it
only ran `make` rather than `ginkgo -r` to run the tests. This was
because the relevant `.travis.yml` entry is `script` rather than the
`scripts` that was being used. This commit fixes that and also adopts
the new `make test` target for running the tests.

[1] https://travis-ci.org/alphagov/paas-metric-exporter/jobs/314763478