Automatically add dot suffix to STATSD_PREFIX opt
StatsD metric names are in a dot seperated notation. It previously was
not clear that you should add a trailing dot when setting a custom
prefix via the STATSD_PREFIX or --statsd-prefix options.

To make for a better user experience, manually adding the trailing dot
is no longer required. It will be added automatically if required.