introduce Sender interface
This introduces a Sender interface, as a point of extension where metrics
systems other than statsd can be introduced.  As part of this, the
metric templating has been pushed into the statsd implementation of this
interface, because that won't be required for prometheus, which uses a
label-based metric naming scheme.