Add prometheus manifest file
This adds a separate manifest that can be pushed using
`cf push -f manifest-prometheus.yml`.

It's not easy to create a prometheus exporter using the existing
manifest, because as well as calling `cf set-env` to enable prometheus
and disable statsd, you also have to bind a route to it and set up a
health check because the manifest specifies "no-route: true" and
"health-check-type: false".  It's much easier to have a separate
manifest that doesn't specify these options so that you get your route
and health check for free on initial push.