Do not include CellID or Job name in metric
Information like the CellID and the Job name (i.e. "diego-cell") is
not unique for a metric. The same application instance can run
in multiple different cells in its lifecycle due restarts of the
app or evictions during platform deployments.

Having this variable metric causes some issues, as metrics
duplications as the app instance is migrated. Specifically, in the
case of prometheus, this service caches the metric based on
the labels, including the cell ID, so we display multiple ghost metrics
metrics from the latest metric from a given cell.

One solution could be refactor the prometheus connector to expire
the metrics after some time (e.g. 30s). But as the CellID and Job
provide little or no information to the tenants, we decided to drop
it.

This has some consequences and gotchas:

 * The statsd metric name does change
 * Prometheus don't get these labels anymore.
 * We still will have ghost metrics with the prometheus exporter:
    * A deleted app would be reported all the time.
    * A renamed app would have duplicated (problem in
      blue-green deploys)
    * A scaled down app (e.g. 4 to 2 instances) will still report
      old instances 3 and 4.

To fix this properly we should still implement some short of
expiration of the cached metrics of the paas-metrics-exporter.

Other problem is that the application caches the metrics forever,
causing effectively a memory leak . This is worse when there are a lot
of deployments that change a lot the GUID of the app.