Allow inject the prometheus.Registerer
When testing our sender it will be convenient we use a dedicated
registry per test, to avoid collisions in the metrics.

This is causing problems when extending the tests, because in the
senders/prometheus_test.go we create one PrometheusSender per test,
but all of them use the prometheus.defaultRegistry.

In this scenario, if we have metrics with the same name, the
PrometheusSenders would try to register duplicates.