Add MetricsExpirer helper to delete old metrics
The prometheus go client metrics vectors objects[1] like CounterVec,
GaugeVec and HistogramVec do not include logic to expire metrics.

This is causing us trouble, as we are efectivelly caching for ever
every single metric that the service receives.

To fix this, we add a MetricsExpirer helper, which would keep
track of the timestamp of the last time a metric, identified
by name and labels, has been seen.

The MetricsExporter would start a goroutine that would iterate across
all the seen metrics, and call the provided callback for those than
are older than a given TTL.

We add the library github.com/mitchellh/hashstructure to quickly
create a hash of the name+labels to keep track the last seen metric.

[1] https://godoc.org/github.com/prometheus/client_golang/prometheus#CounterVec