Expire metrics in prometheus exporter
The prometheus go client metrics vectors objects[1] like CounterVec,
GaugeVec and HistogramVec do not include logic to expire metrics.

This is causing us trouble, as we are efectivelly caching for ever
every single metric that the service receives.

Here we associate a MetricExpirer helper to each one of the metrics
vectors stored in the Prometheus sender. This helper will call
a function to delete the old metrics in the vector if that
name-label combination has not been seen after the given TTL.

We add a new configuration parameter: Prometheus-metric-ttl or
PROMETHEUS_METRIC_TTL to configure the maximum time the metric will
live.

By adding this expire logic, we solve several problems:

 - No more ghost metrics caused by rename of apps, restarts (change
   of cell_id), redeployments, etc
 - Stop leaking memory due endless metric caching.

Note that this solution is suboptimal. The prometheus commutity
discorages this kind of setup, and instead encourages to use the
https://github.com/prometheus/pushgateway or that each application
instance exports their own metrics.  This is a special case where
the paas-metrics-exporter is exposing metrics on behalf of the
different application instances.

But we decided to workaround the issue until we provide a better
alternative.

[1] https://godoc.org/github.com/prometheus/client_golang/prometheus#CounterVec