Add Loggregator sender to the app
We want to start sending these metrics to Loggregator, but this app in
its current state is primarily intended for sending to a StatsD
endpoint. Therefore, we disable the Loggregator sender by default. We
also use a default config with hardcoded values, because we don't have
any integration tests set up and we expect this to be run as a Bosh
release against a real Loggregator.