Add a locking mechanism to the app
We want to be able to run the Metric Exporter in a highly available
manner.  This adds a configuration for running in a locking mode,
whereby Locket is used as part of an `ifrit` process group. If the lock
is lost, or the process dies, another instance can take over exporting
metrics.