explicitly delete metrics when apps are removed
This commit sets things up so that prometheus metrics are explicitly
deleted when apps are deleted.

The approach is to add two new methods to the Sender interface,
AppCreated(guid string) and AppDeleted(guid string), and to set up the
Fetcher to send messages when it starts listening to events for a new
app, and when it stops.

Then, in the Prometheus sender, we split our MetricVecs up by app
guid, so that we can reset all metrics for an app at once, without
having to keep track of each separate metric and labelset.

There are some known drawbacks with this approach, though.  The main
issue is that a set of metrics is scoped to an app *instance*, not
merely to an app.  Therefore, we ought to delete metrics for expired
instances when an app scales down, not just when the whole app is
deleted.

There are some other future improvements which I didn't consider here:

 - pre-register all metrics when an app or instance is created, so
   that counters start from 0 properly
 - move labels like cell, space, and organization to an _info metric [1]

[1]: https://www.robustperception.io/exposing-the-software-version-to-prometheus